wip: check if own bounty before edit/delete

+++ packages/react-app/src/pages/api/bounties/[id].ts
@@ -5,11 +5,14 @@ import { internalServerError, notFound } from '../../../errors';
 import * as service from '../../../services/bounty.service';
 import middlewares from '../../../middlewares';
 import { RoleRestrictions } from '@app/types/Role';
+import { getSession } from 'next-auth/react';
+import { SessionWithToken } from '@app/types/SessionExtended';
 
 const restrictions: RoleRestrictions = {
 	PATCH: ['admin', 'edit-bounties', 'edit-own-bounty'],
 	DELETE: ['admin', 'delete-bounties', 'delete-own-bounty'],
 };
+
 
 export const handler = async (
 	req: NextApiRequest,
@@ -27,9 +30,11 @@ export const handler = async (
 	await dbConnect();
 
 	const bounty = await service.getBounty(id as string);
+
 	if (!bounty) {
 		return notFound(res);
 	}
+
 
 	switch (req.method) {
 	case 'GET':
@@ -40,11 +45,12 @@ export const handler = async (
 	case 'PATCH' :
 		/* Edit a model by its ID */
 		try {
-			const bountyIsEditable = service.canBeEdited({ bounty });
+			const session = await getSession() as SessionWithToken;
+			const bountyIsEditable = service.canBeEdited({ bounty, session, del: false });
 			if (!bountyIsEditable) {
-				return res.status(400).json({
+				return res.status(403).json({
 					success: false,
-					message: 'Unable to edit bounty, as is not in an editable status',
+					message: 'Unauthorized to edit this bounty',
 					bountyStatus: bounty.status,
 				});
 			}
@@ -57,6 +63,15 @@ export const handler = async (
 		
 	case 'DELETE':
 		try {
+			const session = await getSession() as SessionWithToken;
+			const bountyIsDeletable = service.canBeEdited({ bounty, session, del: true });
+			if (!bountyIsDeletable) {
+				return res.status(403).json({
+					success: false,
+					message: 'Unauthorized to delete this bounty',
+					bountyStatus: bounty.status,
+				});
+			}			
 			await service.deleteBounty(id as string);
 			res.status(204).end();
 		} catch (error) {
